backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
duckdb-engine==0.9.2
pandas==2.2.0
aiohttp==3.9.3
python-multipart==0.0.9
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5 